requirements.txt
fastapi~=0.128.0
uvicorn~=0.40.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
greenlet>=3.0.0
python-dotenv>=1.0.0
arq>=0.26.0
redis>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
anyio~=4.12.1